Move user presenters to user folder
