+++ app/Models/User/BigBrother/BigBrother.php
@@ -4,8 +4,8 @@ namespace App\Models\User\BigBrother;
 
 use App\Models\Batch\Batch;
 use App\Models\User\User;
-use App\Orchid\Presenters\BigBrotherPresenter;
-use App\Orchid\Presenters\UserPresenter;
+use App\Orchid\Presenters\User\BigBrotherPresenter;
+use App\Orchid\Presenters\User\UserPresenter;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
