+++ app/Models/User/Farmer/Farmer.php
@@ -7,8 +7,8 @@ use App\Models\Batch\BatchSeedAllocation;
 use App\Models\FarmerReport\FarmerReport;
 use App\Models\Farmland\Farmland;
 use App\Models\User\User;
-use App\Orchid\Presenters\FarmerPresenter;
-use App\Orchid\Presenters\UserPresenter;
+use App\Orchid\Presenters\User\FarmerPresenter;
+use App\Orchid\Presenters\User\UserPresenter;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
