+++ app/Orchid/Presenters/User/FarmerPresenter.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Orchid\Presenters;
+namespace App\Orchid\Presenters\User;
 
 class FarmerPresenter extends UserPresenter
 {
