+++ app/Orchid/Presenters/User/UserPresenter.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Orchid\Presenters;
+namespace App\Orchid\Presenters\User;
 
 use Laravel\Scout\Builder;
 use Orchid\Screen\Contracts\Personable;
